feat: update vue playground
