+++ playground/vue/components/AttachPropsDemo.vue
