+++ playground/vue/components/BuggyComponent.vue
