+++ playground/vue/components/LifecycleErrorComponent.vue
