+++ playground/vue/components/WarnTrigger.vue
