+++ playground/vue/index.html
@@ -3,16 +3,14 @@
 <head>
     <meta charset="UTF-8" />
     <meta content="width=device-width, initial-scale=1.0" name="viewport" />
-    <title>Vanilla JS - Flare Playground</title>
+    <title>Vue - Flare Playground</title>
     <link href="/shared/styles.css" rel="stylesheet" />
 </head>
 <body class="grid grid-cols-[14rem_1fr] min-h-screen bg-gray-50 text-gray-900">
 <div class="relative bg-white border-r border-gray-200" data-active="vue" data-slot="sidebar"></div>
 <main class="p-10 space-y-4">
     <h1 class="text-2xl font-bold">Vue</h1>
-    <div id="root"
-         class="rounded-lg border border-gray-200 bg-white p-4 flex flex-wrap justify-start items-center gap-4">
-    </div>
+    <div id="root"></div>
 </main>
 <script src="/vue/main.ts" type="module"></script>
 </body>
