+++ playground/vue/sections/EnrichmentSection.vue
@@ -1,0 +1,44 @@
+<script lang="ts" setup>
+import { flare } from '../../shared/initFlare';
+
+import Button from '../components/Button.vue';
+import TestSection from '../components/TestSection.vue';
+</script>
+
+<template>
+    <TestSection
+        title="Enrichment: glows and custom context"
+        description="Attaches breadcrumbs (glows) or custom context to a report before sending. Verifies both are forwarded alongside the error."
+        body-class="flex flex-wrap items-center gap-3"
+    >
+        <Button
+            @click="
+                () => {
+                    console.log('Adding glows then reporting error');
+                    flare.glow('User clicked checkout', 'info', { page: '/checkout' });
+                    flare.glow('Payment form submitted', 'info', { method: 'credit_card' });
+                    flare.glow('Payment API responded', 'error', { status: 500 });
+                    flare.report(new Error('Payment processing failed'));
+                }
+            "
+        >
+            Error with glows
+        </Button>
+        <Button
+            @click="
+                () => {
+                    console.log('Adding custom context then reporting error');
+                    flare.addContext('user_id', 'usr_12345');
+                    flare.addContext('plan', 'pro');
+                    flare.addContextGroup('feature_flags', {
+                        new_checkout: true,
+                        dark_mode: false,
+                    });
+                    flare.report(new Error('Error with custom context attached'));
+                }
+            "
+        >
+            Error with custom context
+        </Button>
+    </TestSection>
+</template>
